Update template to parse objects from design
Previously, if a configuration-instantiated object had
children that were instantiated from design, Cacophony
would not generate any code to parse those children.
This commit updates the `designToConfigParser` template
to parse design-instantiated children of
configuration-instantiated objects.

It does so by separating the XML-dependent part of
DPE creation from the XML-independent part, and parsing
the design for the XML-independent part.